--- src/App/App.css
@@ -1,56 +1,0 @@
-.Home {
-  font-family: 'Lato', sans-serif;
-}
-
-.App-logo {
-  height: 25vmin;
-  pointer-events: all;
-  padding-left: 5%;
-}
-
-/* @media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-} */
-
-.App-header {
-  background-color: #FFFFFF;
-  min-height: 10vh;
-  display: flex;
-  flex-direction: row;
-  align-items: left;
-  justify-content: space-between;
-  padding: 3%;
-  font-size: calc(2px + 1.5vmin);
-  color: rgb(2, 34, 2);
-}
-
-.date-container {
-  display: flex;
-  flex-direction: column;
-}
-
-.today {
-  color: #010101;
-}
-
-h2 {
-  text-align: center;
-  margin-top: .5%;
-}
-
-main {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-}
-
-/* @keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-} */
