+++ src/App/App.js
@@ -1,24 +1,14 @@
 import React, { useState, useEffect } from "react";
 import { Routes, Route, Link } from "react-router-dom";
 
-import genlogo from "../assets/genlogo.svg";
+
 import "./App.css";
 import { world } from "../api-calls/api-calls";
 import Card from "../Card/Card";
 import Home from "../Home/Home";
 import StoryDetail from "../StoryDetail/StoryDetail";
 
-const date = new Date();
-const headline = {
-  weekday: "long",
-  year: "numeric",
-  month: "long",
-  day: "numeric",
-};
 
-const todayQuery = date.toISOString().split("T")[0];
-
-const today = date.toLocaleDateString("en-US", headline);
 
 function App() {
   const [worldNews, setWorldNews] = useState([]);
@@ -46,26 +36,17 @@ function App() {
   }
 
   return (
-    <div className="Home">
-      <header className="App-header">
-        <img src={genlogo} className="App-logo" alt="green-energy-news-logo" />
-        <div className="date-container">
-          <h1 className="today">{today}</h1>
-        </div>
-      </header>
-      <h2>Recent Headlines</h2>
-      <main>
-        {worldNews.map((article, index) => (
-          <Card key={index} article={article} />
-        ))}
+      <div>
+        
         <Routes>
-          <Route path="/" element={<Home />} />
+          <Route 
+            path="/" 
+            element={<Home worldNews={worldNews}/>} />
           <Route
-            path="/{article.id}}/{article.title}"
-            element={<StoryDetail />}
-          />
+            path="/:id/:title"
+            element={<StoryDetail worldNews={worldNews}/>} />
         </Routes>
-      </main>
+     
     </div>
   );
 }
