+++ src/StoryDetail/StoryDetail.css
@@ -1,1 +1,24 @@
+.story-detail {
+  font-family: 'Lato', sans-serif;
+  display: flex;
+  flex-direction: column;
+  align-items: left;
+  justify-content: left;
+  width: 60%;
+  margin: 5%;
+  box-sizing: Border-box;
+}
+
+.story-detail img {
+  width: 100%;
+  height: auto;
+  object-fit: cover;
+  max-height: 320px;
+}
+
+.story-detail h4 {
+  font-weight: 600;
+  color: rgb(44, 44, 117);
+}
+
