+++ src/api-calls/api-calls.js
@@ -1,6 +1,7 @@
 import { v4 as uuidv4 } from 'uuid';
 
-
+const todaysDate = new Date()
+const todayQuery = todaysDate.toISOString().split("T")[0];
 //World
 export const world = () => {
   return fetch('https://newsapi.org/v2/everything?q=energy%20And%20clean&domains=washingtonpost.com,bbc.co.uk&apiKey=ec4aaf5184104872884ca453b6fe31b1')
